% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{AGPlot}
\alias{AGPlot}
\title{AGPlot Class}
\description{
The main plot class that handles all visualizations
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AGPlot-new}{\code{AGPlot$new()}}
\item \href{#method-AGPlot-add_layer}{\code{AGPlot$add_layer()}}
\item \href{#method-AGPlot-add_scale}{\code{AGPlot$add_scale()}}
\item \href{#method-AGPlot-set_theme}{\code{AGPlot$set_theme()}}
\item \href{#method-AGPlot-set_facet}{\code{AGPlot$set_facet()}}
\item \href{#method-AGPlot-set_coord}{\code{AGPlot$set_coord()}}
\item \href{#method-AGPlot-add_annotation}{\code{AGPlot$add_annotation()}}
\item \href{#method-AGPlot-compute_ranges}{\code{AGPlot$compute_ranges()}}
\item \href{#method-AGPlot-collect_legends}{\code{AGPlot$collect_legends()}}
\item \href{#method-AGPlot-plot}{\code{AGPlot$plot()}}
\item \href{#method-AGPlot-clone}{\code{AGPlot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AGPlot-new"></a>}}
\if{latex}{\out{\hypertarget{method-AGPlot-new}{}}}
\subsection{Method \code{new()}}{
Create a new plot object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AGPlot$new(data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A data frame containing the data to plot}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AGPlot-add_layer"></a>}}
\if{latex}{\out{\hypertarget{method-AGPlot-add_layer}{}}}
\subsection{Method \code{add_layer()}}{
Add a new layer to the plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AGPlot$add_layer(layer)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{layer}}{A layer object to add to the plot}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AGPlot-add_scale"></a>}}
\if{latex}{\out{\hypertarget{method-AGPlot-add_scale}{}}}
\subsection{Method \code{add_scale()}}{
Add a scale to the plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AGPlot$add_scale(scale)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scale}}{A scale object to add to the plot}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AGPlot-set_theme"></a>}}
\if{latex}{\out{\hypertarget{method-AGPlot-set_theme}{}}}
\subsection{Method \code{set_theme()}}{
Set the theme for the plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AGPlot$set_theme(theme)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{theme}}{A theme object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AGPlot-set_facet"></a>}}
\if{latex}{\out{\hypertarget{method-AGPlot-set_facet}{}}}
\subsection{Method \code{set_facet()}}{
Set the faceting specification for the plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AGPlot$set_facet(facet)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{facet}}{A facet object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AGPlot-set_coord"></a>}}
\if{latex}{\out{\hypertarget{method-AGPlot-set_coord}{}}}
\subsection{Method \code{set_coord()}}{
Set the coordinate system for the plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AGPlot$set_coord(coord)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{coord}}{A coordinate system object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AGPlot-add_annotation"></a>}}
\if{latex}{\out{\hypertarget{method-AGPlot-add_annotation}{}}}
\subsection{Method \code{add_annotation()}}{
Add an annotation to the plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AGPlot$add_annotation(annotation)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{annotation}}{An annotation object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AGPlot-compute_ranges"></a>}}
\if{latex}{\out{\hypertarget{method-AGPlot-compute_ranges}{}}}
\subsection{Method \code{compute_ranges()}}{
Calculate plot ranges from data and layers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AGPlot$compute_ranges()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AGPlot-collect_legends"></a>}}
\if{latex}{\out{\hypertarget{method-AGPlot-collect_legends}{}}}
\subsection{Method \code{collect_legends()}}{
Collect legends from layers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AGPlot$collect_legends()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AGPlot-plot"></a>}}
\if{latex}{\out{\hypertarget{method-AGPlot-plot}{}}}
\subsection{Method \code{plot()}}{
Render the plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AGPlot$plot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AGPlot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AGPlot-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AGPlot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
