% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation.R
\name{animate_time_series}
\alias{animate_time_series}
\title{Create a time series animation}
\usage{
animate_time_series(time_column, window_size = NULL, duration = 1000)
}
\arguments{
\item{time_column}{Name of time column}

\item{window_size}{Size of moving window}

\item{duration}{Duration in milliseconds}
}
\description{
Create a time series animation
}
