% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation.R
\name{animate_transition}
\alias{animate_transition}
\title{Create a transition animation}
\usage{
animate_transition(states, duration = 1000, easing = "linear")
}
\arguments{
\item{states}{List of data frames representing states}

\item{duration}{Duration in milliseconds}

\item{easing}{Easing function name}
}
\description{
Create a transition animation
}
