% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotate_arrow}
\alias{annotate_arrow}
\title{Create an arrow annotation}
\usage{
annotate_arrow(x0, y0, x1, y1, color = "black", size = 1, arrow_size = 0.1)
}
\arguments{
\item{x0}{Start x coordinate}

\item{y0}{Start y coordinate}

\item{x1}{End x coordinate}

\item{y1}{End y coordinate}

\item{color}{Arrow color}

\item{size}{Line size}

\item{arrow_size}{Size of arrow head}
}
\description{
Create an arrow annotation
}
