% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotate_reference_line}
\alias{annotate_reference_line}
\title{Create a reference line annotation}
\usage{
annotate_reference_line(
  value,
  orientation = "h",
  color = "gray50",
  size = 1,
  linetype = "dashed"
)
}
\arguments{
\item{value}{Position of the line}

\item{orientation}{"h" for horizontal or "v" for vertical}

\item{color}{Line color}

\item{size}{Line size}

\item{linetype}{Line type}
}
\description{
Create a reference line annotation
}
