% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotate_text}
\alias{annotate_text}
\title{Create a text annotation}
\usage{
annotate_text(x, y, label, color = "black", size = 10, angle = 0)
}
\arguments{
\item{x}{x coordinate}

\item{y}{y coordinate}

\item{label}{Text to display}

\item{color}{Text color}

\item{size}{Text size}

\item{angle}{Rotation angle in degrees}
}
\description{
Create a text annotation
}
