% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{Annotation}
\alias{Annotation}
\title{Annotation Class}
\description{
Base class for all annotations
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Annotation-new}{\code{Annotation$new()}}
\item \href{#method-Annotation-render}{\code{Annotation$render()}}
\item \href{#method-Annotation-clone}{\code{Annotation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Annotation-new"></a>}}
\if{latex}{\out{\hypertarget{method-Annotation-new}{}}}
\subsection{Method \code{new()}}{
Initialize annotation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Annotation$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Annotation-render"></a>}}
\if{latex}{\out{\hypertarget{method-Annotation-render}{}}}
\subsection{Method \code{render()}}{
Render annotation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Annotation$render(viewport)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{viewport}}{The current viewport}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Annotation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Annotation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Annotation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
