% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{AnnotationReferenceLine}
\alias{AnnotationReferenceLine}
\title{Reference Line Annotation}
\description{
Add reference lines to the plot
}
\section{Super class}{
\code{\link[AdvancedGgplot:Annotation]{AdvancedGgplot::Annotation}} -> \code{AnnotationReferenceLine}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AnnotationReferenceLine-new}{\code{AnnotationReferenceLine$new()}}
\item \href{#method-AnnotationReferenceLine-render}{\code{AnnotationReferenceLine$render()}}
\item \href{#method-AnnotationReferenceLine-clone}{\code{AnnotationReferenceLine$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotationReferenceLine-new"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotationReferenceLine-new}{}}}
\subsection{Method \code{new()}}{
Initialize reference line annotation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationReferenceLine$new(
  value,
  orientation = "h",
  color = "gray50",
  size = 1,
  linetype = "dashed"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Position of the line}

\item{\code{orientation}}{"h" for horizontal or "v" for vertical}

\item{\code{color}}{Line color}

\item{\code{size}}{Line size}

\item{\code{linetype}}{Line type ("solid", "dashed", "dotted", etc.)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotationReferenceLine-render"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotationReferenceLine-render}{}}}
\subsection{Method \code{render()}}{
Render reference line annotation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationReferenceLine$render(viewport)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{viewport}}{The current viewport}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotationReferenceLine-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotationReferenceLine-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationReferenceLine$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
