% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{AnnotationText}
\alias{AnnotationText}
\title{Text Annotation}
\description{
Add text labels to the plot
}
\section{Super class}{
\code{\link[AdvancedGgplot:Annotation]{AdvancedGgplot::Annotation}} -> \code{AnnotationText}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AnnotationText-new}{\code{AnnotationText$new()}}
\item \href{#method-AnnotationText-render}{\code{AnnotationText$render()}}
\item \href{#method-AnnotationText-clone}{\code{AnnotationText$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotationText-new"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotationText-new}{}}}
\subsection{Method \code{new()}}{
Initialize text annotation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationText$new(x, y, label, color = "black", size = 10, angle = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{x coordinate}

\item{\code{y}}{y coordinate}

\item{\code{label}}{Text to display}

\item{\code{color}}{Text color}

\item{\code{size}}{Text size}

\item{\code{angle}}{Rotation angle in degrees}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotationText-render"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotationText-render}{}}}
\subsection{Method \code{render()}}{
Render text annotation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationText$render(viewport)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{viewport}}{The current viewport}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotationText-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotationText-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationText$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
