% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{arrange_plots}
\alias{arrange_plots}
\title{Create a new layout}
\usage{
arrange_plots(..., ncol = NULL, nrow = NULL, widths = NULL, heights = NULL)
}
\arguments{
\item{...}{Plots to arrange}

\item{ncol}{Number of columns}

\item{nrow}{Number of rows}

\item{widths}{Relative widths of columns}

\item{heights}{Relative heights of rows}
}
\description{
Create a new layout
}
