% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{ColorPalette}
\alias{ColorPalette}
\title{Color Palette Class}
\description{
Handles color generation and palette management
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ColorPalette-new}{\code{ColorPalette$new()}}
\item \href{#method-ColorPalette-get_colors}{\code{ColorPalette$get_colors()}}
\item \href{#method-ColorPalette-clone}{\code{ColorPalette$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ColorPalette-new"></a>}}
\if{latex}{\out{\hypertarget{method-ColorPalette-new}{}}}
\subsection{Method \code{new()}}{
Initialize color palette
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorPalette$new(colors, interpolate = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{colors}}{Vector of colors}

\item{\code{interpolate}}{Whether to interpolate between colors}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ColorPalette-get_colors"></a>}}
\if{latex}{\out{\hypertarget{method-ColorPalette-get_colors}{}}}
\subsection{Method \code{get_colors()}}{
Get colors for n values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorPalette$get_colors(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of colors needed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ColorPalette-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ColorPalette-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorPalette$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
