% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{ColorScale}
\alias{ColorScale}
\title{Color Scale Class}
\description{
Maps data values to colors
}
\section{Super class}{
\code{\link[AdvancedGgplot:Scale]{AdvancedGgplot::Scale}} -> \code{ColorScale}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ColorScale-new}{\code{ColorScale$new()}}
\item \href{#method-ColorScale-transform}{\code{ColorScale$transform()}}
\item \href{#method-ColorScale-clone}{\code{ColorScale$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Scale" data-id="render_axis"><a href='../../AdvancedGgplot/html/Scale.html#method-Scale-render_axis'><code>AdvancedGgplot::Scale$render_axis()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ColorScale-new"></a>}}
\if{latex}{\out{\hypertarget{method-ColorScale-new}{}}}
\subsection{Method \code{new()}}{
Initialize color scale
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorScale$new(aesthetic, palette, na_color = "#808080")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{aesthetic}}{The aesthetic to scale}

\item{\code{palette}}{Color palette to use}

\item{\code{na_color}}{Color for NA values}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ColorScale-transform"></a>}}
\if{latex}{\out{\hypertarget{method-ColorScale-transform}{}}}
\subsection{Method \code{transform()}}{
Transform data values to colors
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorScale$transform(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Values to transform}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ColorScale-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ColorScale-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorScale$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
