% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{Coord}
\alias{Coord}
\title{Coordinate System Class}
\description{
Base class for coordinate systems
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Coord-transform}{\code{Coord$transform()}}
\item \href{#method-Coord-transform_limits}{\code{Coord$transform_limits()}}
\item \href{#method-Coord-clone}{\code{Coord$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Coord-transform"></a>}}
\if{latex}{\out{\hypertarget{method-Coord-transform}{}}}
\subsection{Method \code{transform()}}{
Transform coordinates
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Coord$transform(x, y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{x-coordinates}

\item{\code{y}}{y-coordinates}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Coord-transform_limits"></a>}}
\if{latex}{\out{\hypertarget{method-Coord-transform_limits}{}}}
\subsection{Method \code{transform_limits()}}{
Transform limits
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Coord$transform_limits(xlim, ylim)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xlim}}{x-axis limits}

\item{\code{ylim}}{y-axis limits}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Coord-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Coord-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Coord$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
