% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{coord_cartesian}
\alias{coord_cartesian}
\title{Create Cartesian coordinates}
\usage{
coord_cartesian(xlim = NULL, ylim = NULL, expand = TRUE)
}
\arguments{
\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}

\item{expand}{Whether to expand limits}
}
\description{
Create Cartesian coordinates
}
