% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{coord_polar}
\alias{coord_polar}
\title{Create polar coordinates}
\usage{
coord_polar(theta = "x", start = 0, direction = 1)
}
\arguments{
\item{theta}{Variable to map to angle}

\item{start}{Starting angle}

\item{direction}{Direction of rotation}
}
\description{
Create polar coordinates
}
