% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{CoordCartesian}
\alias{CoordCartesian}
\title{Cartesian Coordinate System}
\description{
Standard Cartesian coordinates
}
\section{Super class}{
\code{\link[AdvancedGgplot:Coord]{AdvancedGgplot::Coord}} -> \code{CoordCartesian}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CoordCartesian-new}{\code{CoordCartesian$new()}}
\item \href{#method-CoordCartesian-clone}{\code{CoordCartesian$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Coord" data-id="transform"><a href='../../AdvancedGgplot/html/Coord.html#method-Coord-transform'><code>AdvancedGgplot::Coord$transform()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Coord" data-id="transform_limits"><a href='../../AdvancedGgplot/html/Coord.html#method-Coord-transform_limits'><code>AdvancedGgplot::Coord$transform_limits()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordCartesian-new"></a>}}
\if{latex}{\out{\hypertarget{method-CoordCartesian-new}{}}}
\subsection{Method \code{new()}}{
Initialize Cartesian coordinates
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordCartesian$new(xlim = NULL, ylim = NULL, expand = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xlim}}{x-axis limits}

\item{\code{ylim}}{y-axis limits}

\item{\code{expand}}{Whether to expand limits}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordCartesian-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CoordCartesian-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordCartesian$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
