% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{CoordLog}
\alias{CoordLog}
\title{Log Coordinate System}
\description{
Logarithmic coordinate transformation
}
\section{Super class}{
\code{\link[AdvancedGgplot:Coord]{AdvancedGgplot::Coord}} -> \code{CoordLog}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CoordLog-new}{\code{CoordLog$new()}}
\item \href{#method-CoordLog-transform}{\code{CoordLog$transform()}}
\item \href{#method-CoordLog-transform_limits}{\code{CoordLog$transform_limits()}}
\item \href{#method-CoordLog-clone}{\code{CoordLog$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordLog-new"></a>}}
\if{latex}{\out{\hypertarget{method-CoordLog-new}{}}}
\subsection{Method \code{new()}}{
Initialize log coordinates
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordLog$new(base = exp(1))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{base}}{Base of logarithm}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordLog-transform"></a>}}
\if{latex}{\out{\hypertarget{method-CoordLog-transform}{}}}
\subsection{Method \code{transform()}}{
Transform coordinates to log scale
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordLog$transform(x, y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{x-coordinates}

\item{\code{y}}{y-coordinates}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordLog-transform_limits"></a>}}
\if{latex}{\out{\hypertarget{method-CoordLog-transform_limits}{}}}
\subsection{Method \code{transform_limits()}}{
Transform limits to log scale
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordLog$transform_limits(xlim, ylim)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xlim}}{x-axis limits}

\item{\code{ylim}}{y-axis limits}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordLog-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CoordLog-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordLog$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
