% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{CoordPolar}
\alias{CoordPolar}
\title{Polar Coordinate System}
\description{
Polar coordinates for circular plots
}
\section{Super class}{
\code{\link[AdvancedGgplot:Coord]{AdvancedGgplot::Coord}} -> \code{CoordPolar}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CoordPolar-new}{\code{CoordPolar$new()}}
\item \href{#method-CoordPolar-transform}{\code{CoordPolar$transform()}}
\item \href{#method-CoordPolar-clone}{\code{CoordPolar$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Coord" data-id="transform_limits"><a href='../../AdvancedGgplot/html/Coord.html#method-Coord-transform_limits'><code>AdvancedGgplot::Coord$transform_limits()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordPolar-new"></a>}}
\if{latex}{\out{\hypertarget{method-CoordPolar-new}{}}}
\subsection{Method \code{new()}}{
Initialize polar coordinates
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordPolar$new(theta = "x", start = 0, direction = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{theta}}{Variable to map to angle ("x" or "y")}

\item{\code{start}}{Starting angle in radians}

\item{\code{direction}}{Direction of rotation (1 or -1)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordPolar-transform"></a>}}
\if{latex}{\out{\hypertarget{method-CoordPolar-transform}{}}}
\subsection{Method \code{transform()}}{
Transform coordinates to polar
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordPolar$transform(x, y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{x-coordinates}

\item{\code{y}}{y-coordinates}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordPolar-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CoordPolar-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordPolar$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
