% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{create_legend}
\alias{create_legend}
\title{Create a new legend}
\usage{
create_legend(aesthetic, title = NULL, values = NULL, labels = NULL)
}
\arguments{
\item{aesthetic}{The aesthetic this legend represents}

\item{title}{Legend title}

\item{values}{The values to show in the legend}

\item{labels}{Labels for the values}
}
\description{
Create a new legend
}
