% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet.R
\name{facet_grid}
\alias{facet_grid}
\title{Create a grid faceting specification}
\usage{
facet_grid(rows = NULL, cols = NULL, scales = "fixed")
}
\arguments{
\item{rows}{Variable(s) for row facets}

\item{cols}{Variable(s) for column facets}

\item{scales}{Scale handling}
}
\description{
Create a grid faceting specification
}
