% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet.R
\name{facet_wrap}
\alias{facet_wrap}
\title{Create a wrapped faceting specification}
\usage{
facet_wrap(facets, ncol = NULL, nrow = NULL, scales = "fixed")
}
\arguments{
\item{facets}{Variable(s) to facet by}

\item{ncol}{Number of columns}

\item{nrow}{Number of rows}

\item{scales}{Scale handling}
}
\description{
Create a wrapped faceting specification
}
