% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet.R
\name{FacetGrid}
\alias{FacetGrid}
\title{Grid Facet Class}
\description{
Create grid-based faceting
}
\section{Super class}{
\code{\link[AdvancedGgplot:Facet]{AdvancedGgplot::Facet}} -> \code{FacetGrid}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FacetGrid-render}{\code{FacetGrid$render()}}
\item \href{#method-FacetGrid-clone}{\code{FacetGrid$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Facet" data-id="generate_scales"><a href='../../AdvancedGgplot/html/Facet.html#method-Facet-generate_scales'><code>AdvancedGgplot::Facet$generate_scales()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Facet" data-id="get_dimensions"><a href='../../AdvancedGgplot/html/Facet.html#method-Facet-get_dimensions'><code>AdvancedGgplot::Facet$get_dimensions()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Facet" data-id="initialize"><a href='../../AdvancedGgplot/html/Facet.html#method-Facet-initialize'><code>AdvancedGgplot::Facet$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Facet" data-id="split_data"><a href='../../AdvancedGgplot/html/Facet.html#method-Facet-split_data'><code>AdvancedGgplot::Facet$split_data()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FacetGrid-render"></a>}}
\if{latex}{\out{\hypertarget{method-FacetGrid-render}{}}}
\subsection{Method \code{render()}}{
Render faceted plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FacetGrid$render(plot, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plot}}{Plot object to facet}

\item{\code{data}}{Data frame containing the data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FacetGrid-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FacetGrid-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FacetGrid$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
