% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet.R
\name{FacetWrap}
\alias{FacetWrap}
\title{Wrap Facet Class}
\description{
Create wrapped faceting
}
\section{Super class}{
\code{\link[AdvancedGgplot:Facet]{AdvancedGgplot::Facet}} -> \code{FacetWrap}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FacetWrap-new}{\code{FacetWrap$new()}}
\item \href{#method-FacetWrap-get_dimensions}{\code{FacetWrap$get_dimensions()}}
\item \href{#method-FacetWrap-clone}{\code{FacetWrap$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Facet" data-id="generate_scales"><a href='../../AdvancedGgplot/html/Facet.html#method-Facet-generate_scales'><code>AdvancedGgplot::Facet$generate_scales()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Facet" data-id="split_data"><a href='../../AdvancedGgplot/html/Facet.html#method-Facet-split_data'><code>AdvancedGgplot::Facet$split_data()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FacetWrap-new"></a>}}
\if{latex}{\out{\hypertarget{method-FacetWrap-new}{}}}
\subsection{Method \code{new()}}{
Initialize wrapped faceting
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FacetWrap$new(facets, ncol = NULL, nrow = NULL, scales = "fixed")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{facets}}{Variable(s) to facet by}

\item{\code{ncol}}{Number of columns}

\item{\code{nrow}}{Number of rows}

\item{\code{scales}}{Scale handling}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FacetWrap-get_dimensions"></a>}}
\if{latex}{\out{\hypertarget{method-FacetWrap-get_dimensions}{}}}
\subsection{Method \code{get_dimensions()}}{
Calculate layout dimensions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FacetWrap$get_dimensions(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FacetWrap-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FacetWrap-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FacetWrap$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
