% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_bar.R
\name{geom_bar}
\alias{geom_bar}
\title{Create a bar layer}
\usage{
geom_bar(mapping = NULL, width = 0.9)
}
\arguments{
\item{mapping}{Aesthetic mappings}

\item{width}{Bar width as a proportion of the resolution of the data}
}
\description{
Create a bar layer
}
