% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_boxplot.R
\name{geom_boxplot}
\alias{geom_boxplot}
\title{Create a boxplot layer}
\usage{
geom_boxplot(mapping = NULL, width = 0.75)
}
\arguments{
\item{mapping}{Aesthetic mappings}

\item{width}{Width of the boxplot relative to the category width}
}
\description{
Create a boxplot layer
}
