% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_contour.R
\name{geom_contour}
\alias{geom_contour}
\title{Create a contour layer}
\usage{
geom_contour(mapping = NULL, bins = 10)
}
\arguments{
\item{mapping}{Aesthetic mappings}

\item{bins}{Number of contour levels}
}
\description{
Create a contour layer
}
