% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_contour.R
\name{geom_contour_filled}
\alias{geom_contour_filled}
\title{Create a filled contour layer}
\usage{
geom_contour_filled(mapping = NULL, bins = 10)
}
\arguments{
\item{mapping}{Aesthetic mappings}

\item{bins}{Number of contour levels}
}
\description{
Create a filled contour layer
}
