% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_density.R
\name{geom_density}
\alias{geom_density}
\title{Create a density plot layer}
\usage{
geom_density(mapping, kernel = "gaussian", bw = "nrd0")
}
\arguments{
\item{mapping}{Aesthetic mappings}

\item{kernel}{Kernel function to use for density estimation}

\item{bw}{Bandwidth method or value}
}
\value{
A GeomDensity object
}
\description{
Create a density plot layer
}
\examples{
\dontrun{
plot <- AGPlot$new(data)
plot$add_layer(geom_density(list(x = "variable")))
}
}
