% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_histogram.R
\name{geom_histogram}
\alias{geom_histogram}
\title{Create a histogram layer}
\usage{
geom_histogram(mapping, bins = NULL)
}
\arguments{
\item{mapping}{Aesthetic mappings}

\item{bins}{Number of bins}
}
\value{
A GeomHistogram object
}
\description{
Create a histogram layer
}
\examples{
\dontrun{
plot <- AGPlot$new(data)
plot$add_layer(geom_histogram(list(x = "variable")))
}
}
