% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_point.R
\name{geom_point}
\alias{geom_point}
\title{Create a point layer}
\usage{
geom_point(mapping = NULL, size = NULL)
}
\arguments{
\item{mapping}{Aesthetic mappings}

\item{size}{Point size (optional)}
}
\value{
A GeomPoint object
}
\description{
Create a point layer
}
\examples{
\dontrun{
plot <- AGPlot$new(data)
plot$add_layer(geom_point(list(x = "x", y = "y")))
}
}
