% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_smooth.R
\name{geom_smooth}
\alias{geom_smooth}
\title{Create a smoothed line layer with confidence intervals}
\usage{
geom_smooth(mapping, method = "loess", span = 0.75)
}
\arguments{
\item{mapping}{Aesthetic mappings}

\item{method}{Smoothing method ("loess" or "lm")}

\item{span}{Parameter controlling the amount of smoothing}
}
\value{
A GeomSmooth object
}
\description{
Create a smoothed line layer with confidence intervals
}
\examples{
\dontrun{
plot <- AGPlot$new(data)
plot$add_layer(geom_smooth(list(x = "x", y = "y")))
}
}
