% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_violin.R
\name{geom_violin}
\alias{geom_violin}
\title{Create a violin plot layer}
\usage{
geom_violin(mapping = NULL, width = 0.8, scale = "area")
}
\arguments{
\item{mapping}{Aesthetic mappings}

\item{width}{Width of the violin plot relative to the category width}

\item{scale}{How to scale the violins: "area", "count", or "width"}
}
\description{
Create a violin plot layer
}
