% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_area.R
\name{GeomArea}
\alias{GeomArea}
\title{Area Geometry}
\description{
Create filled area plots
}
\section{Super class}{
\code{\link[AdvancedGgplot:Layer]{AdvancedGgplot::Layer}} -> \code{GeomArea}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GeomArea-render}{\code{GeomArea$render()}}
\item \href{#method-GeomArea-clone}{\code{GeomArea$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Layer" data-id="eval_aes"><a href='../../AdvancedGgplot/html/Layer.html#method-Layer-eval_aes'><code>AdvancedGgplot::Layer$eval_aes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Layer" data-id="initialize"><a href='../../AdvancedGgplot/html/Layer.html#method-Layer-initialize'><code>AdvancedGgplot::Layer$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeomArea-render"></a>}}
\if{latex}{\out{\hypertarget{method-GeomArea-render}{}}}
\subsection{Method \code{render()}}{
Render area on the plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeomArea$render(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data frame containing the values to plot}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeomArea-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GeomArea-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeomArea$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
