% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_bar.R
\name{GeomBar}
\alias{GeomBar}
\title{Bar Geometry}
\description{
Create bar plots
}
\section{Super class}{
\code{\link[AdvancedGgplot:Layer]{AdvancedGgplot::Layer}} -> \code{GeomBar}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GeomBar-new}{\code{GeomBar$new()}}
\item \href{#method-GeomBar-render}{\code{GeomBar$render()}}
\item \href{#method-GeomBar-clone}{\code{GeomBar$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Layer" data-id="eval_aes"><a href='../../AdvancedGgplot/html/Layer.html#method-Layer-eval_aes'><code>AdvancedGgplot::Layer$eval_aes()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeomBar-new"></a>}}
\if{latex}{\out{\hypertarget{method-GeomBar-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new bar geometry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeomBar$new(mapping = NULL, width = 0.9)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mapping}}{Aesthetic mappings}

\item{\code{width}}{Bar width as a proportion of the resolution of the data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeomBar-render"></a>}}
\if{latex}{\out{\hypertarget{method-GeomBar-render}{}}}
\subsection{Method \code{render()}}{
Render bars on the plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeomBar$render(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data frame containing the values to plot}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeomBar-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GeomBar-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeomBar$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
