% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_boxplot.R
\name{GeomBoxplot}
\alias{GeomBoxplot}
\title{Boxplot Geometry}
\description{
Create a boxplot layer for statistical summaries
}
\section{Super class}{
\code{\link[AdvancedGgplot:Layer]{AdvancedGgplot::Layer}} -> \code{GeomBoxplot}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GeomBoxplot-render}{\code{GeomBoxplot$render()}}
\item \href{#method-GeomBoxplot-clone}{\code{GeomBoxplot$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Layer" data-id="eval_aes"><a href='../../AdvancedGgplot/html/Layer.html#method-Layer-eval_aes'><code>AdvancedGgplot::Layer$eval_aes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Layer" data-id="initialize"><a href='../../AdvancedGgplot/html/Layer.html#method-Layer-initialize'><code>AdvancedGgplot::Layer$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeomBoxplot-render"></a>}}
\if{latex}{\out{\hypertarget{method-GeomBoxplot-render}{}}}
\subsection{Method \code{render()}}{
Render boxplots on the plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeomBoxplot$render(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data frame containing the points to plot}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeomBoxplot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GeomBoxplot-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeomBoxplot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
