% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_contour.R
\name{GeomContour}
\alias{GeomContour}
\title{Contour Geometry}
\description{
Create contour plots
}
\section{Super class}{
\code{\link[AdvancedGgplot:Layer]{AdvancedGgplot::Layer}} -> \code{GeomContour}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GeomContour-new}{\code{GeomContour$new()}}
\item \href{#method-GeomContour-render}{\code{GeomContour$render()}}
\item \href{#method-GeomContour-clone}{\code{GeomContour$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Layer" data-id="eval_aes"><a href='../../AdvancedGgplot/html/Layer.html#method-Layer-eval_aes'><code>AdvancedGgplot::Layer$eval_aes()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeomContour-new"></a>}}
\if{latex}{\out{\hypertarget{method-GeomContour-new}{}}}
\subsection{Method \code{new()}}{
Initialize contour geometry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeomContour$new(mapping = NULL, bins = 10)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mapping}}{Aesthetic mappings}

\item{\code{bins}}{Number of contour levels}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeomContour-render"></a>}}
\if{latex}{\out{\hypertarget{method-GeomContour-render}{}}}
\subsection{Method \code{render()}}{
Render contours on the plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeomContour$render(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data frame containing the values to plot}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeomContour-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GeomContour-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeomContour$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
