% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_contour.R
\name{GeomContourFilled}
\alias{GeomContourFilled}
\title{Filled Contour Geometry}
\description{
Create filled contour plots
}
\section{Super classes}{
\code{\link[AdvancedGgplot:Layer]{AdvancedGgplot::Layer}} -> \code{\link[AdvancedGgplot:GeomContour]{AdvancedGgplot::GeomContour}} -> \code{GeomContourFilled}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GeomContourFilled-render}{\code{GeomContourFilled$render()}}
\item \href{#method-GeomContourFilled-clone}{\code{GeomContourFilled$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Layer" data-id="eval_aes"><a href='../../AdvancedGgplot/html/Layer.html#method-Layer-eval_aes'><code>AdvancedGgplot::Layer$eval_aes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="GeomContour" data-id="initialize"><a href='../../AdvancedGgplot/html/GeomContour.html#method-GeomContour-initialize'><code>AdvancedGgplot::GeomContour$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeomContourFilled-render"></a>}}
\if{latex}{\out{\hypertarget{method-GeomContourFilled-render}{}}}
\subsection{Method \code{render()}}{
Render filled contours on the plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeomContourFilled$render(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data frame containing the values to plot}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeomContourFilled-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GeomContourFilled-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeomContourFilled$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
