% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_density.R
\name{GeomDensity}
\alias{GeomDensity}
\title{Density Plot Layer}
\description{
Creates a density plot layer
}
\section{Super class}{
\code{\link[AdvancedGgplot:Layer]{AdvancedGgplot::Layer}} -> \code{GeomDensity}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GeomDensity-new}{\code{GeomDensity$new()}}
\item \href{#method-GeomDensity-render}{\code{GeomDensity$render()}}
\item \href{#method-GeomDensity-clone}{\code{GeomDensity$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Layer" data-id="eval_aes"><a href='../../AdvancedGgplot/html/Layer.html#method-Layer-eval_aes'><code>AdvancedGgplot::Layer$eval_aes()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeomDensity-new"></a>}}
\if{latex}{\out{\hypertarget{method-GeomDensity-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new density layer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeomDensity$new(mapping, kernel = "gaussian", bw = "nrd0")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mapping}}{Aesthetic mappings}

\item{\code{kernel}}{Kernel function to use for density estimation}

\item{\code{bw}}{Bandwidth method or value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeomDensity-render"></a>}}
\if{latex}{\out{\hypertarget{method-GeomDensity-render}{}}}
\subsection{Method \code{render()}}{
Render the density plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeomDensity$render(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data to render}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeomDensity-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GeomDensity-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeomDensity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
