% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_histogram.R
\name{GeomHistogram}
\alias{GeomHistogram}
\title{Histogram Layer}
\description{
Creates a histogram layer
}
\section{Super class}{
\code{\link[AdvancedGgplot:Layer]{AdvancedGgplot::Layer}} -> \code{GeomHistogram}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GeomHistogram-new}{\code{GeomHistogram$new()}}
\item \href{#method-GeomHistogram-render}{\code{GeomHistogram$render()}}
\item \href{#method-GeomHistogram-clone}{\code{GeomHistogram$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Layer" data-id="eval_aes"><a href='../../AdvancedGgplot/html/Layer.html#method-Layer-eval_aes'><code>AdvancedGgplot::Layer$eval_aes()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeomHistogram-new"></a>}}
\if{latex}{\out{\hypertarget{method-GeomHistogram-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new histogram layer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeomHistogram$new(mapping, bins = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mapping}}{Aesthetic mappings}

\item{\code{bins}}{Number of bins}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeomHistogram-render"></a>}}
\if{latex}{\out{\hypertarget{method-GeomHistogram-render}{}}}
\subsection{Method \code{render()}}{
Render the histogram
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeomHistogram$render(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data to render}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeomHistogram-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GeomHistogram-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeomHistogram$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
