% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_line.R
\name{GeomLine}
\alias{GeomLine}
\title{Line Geometry}
\description{
Create a line plot layer
}
\section{Super class}{
\code{\link[AdvancedGgplot:Layer]{AdvancedGgplot::Layer}} -> \code{GeomLine}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GeomLine-render}{\code{GeomLine$render()}}
\item \href{#method-GeomLine-clone}{\code{GeomLine$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Layer" data-id="eval_aes"><a href='../../AdvancedGgplot/html/Layer.html#method-Layer-eval_aes'><code>AdvancedGgplot::Layer$eval_aes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Layer" data-id="initialize"><a href='../../AdvancedGgplot/html/Layer.html#method-Layer-initialize'><code>AdvancedGgplot::Layer$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeomLine-render"></a>}}
\if{latex}{\out{\hypertarget{method-GeomLine-render}{}}}
\subsection{Method \code{render()}}{
Render lines on the plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeomLine$render(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data frame containing the points to connect}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeomLine-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GeomLine-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeomLine$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
