% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_point.R
\name{GeomPoint}
\alias{GeomPoint}
\title{Point Geometry}
\description{
Create a scatter plot layer
}
\section{Super class}{
\code{\link[AdvancedGgplot:Layer]{AdvancedGgplot::Layer}} -> \code{GeomPoint}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GeomPoint-render}{\code{GeomPoint$render()}}
\item \href{#method-GeomPoint-clone}{\code{GeomPoint$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Layer" data-id="eval_aes"><a href='../../AdvancedGgplot/html/Layer.html#method-Layer-eval_aes'><code>AdvancedGgplot::Layer$eval_aes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Layer" data-id="initialize"><a href='../../AdvancedGgplot/html/Layer.html#method-Layer-initialize'><code>AdvancedGgplot::Layer$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeomPoint-render"></a>}}
\if{latex}{\out{\hypertarget{method-GeomPoint-render}{}}}
\subsection{Method \code{render()}}{
Render points on the plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeomPoint$render(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data frame containing the points to plot}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeomPoint-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GeomPoint-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeomPoint$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
