% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_ribbon.R
\name{GeomRibbon}
\alias{GeomRibbon}
\title{Ribbon Geometry}
\description{
Create ribbon plots (e.g., confidence bands)
}
\section{Super class}{
\code{\link[AdvancedGgplot:Layer]{AdvancedGgplot::Layer}} -> \code{GeomRibbon}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GeomRibbon-render}{\code{GeomRibbon$render()}}
\item \href{#method-GeomRibbon-clone}{\code{GeomRibbon$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Layer" data-id="eval_aes"><a href='../../AdvancedGgplot/html/Layer.html#method-Layer-eval_aes'><code>AdvancedGgplot::Layer$eval_aes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Layer" data-id="initialize"><a href='../../AdvancedGgplot/html/Layer.html#method-Layer-initialize'><code>AdvancedGgplot::Layer$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeomRibbon-render"></a>}}
\if{latex}{\out{\hypertarget{method-GeomRibbon-render}{}}}
\subsection{Method \code{render()}}{
Render ribbon on the plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeomRibbon$render(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data frame containing the values to plot}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeomRibbon-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GeomRibbon-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeomRibbon$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
