% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_smooth.R
\name{GeomSmooth}
\alias{GeomSmooth}
\title{Smoothed Line Layer}
\description{
Creates a smoothed line layer with optional confidence intervals
}
\section{Super class}{
\code{\link[AdvancedGgplot:Layer]{AdvancedGgplot::Layer}} -> \code{GeomSmooth}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GeomSmooth-new}{\code{GeomSmooth$new()}}
\item \href{#method-GeomSmooth-render}{\code{GeomSmooth$render()}}
\item \href{#method-GeomSmooth-clone}{\code{GeomSmooth$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Layer" data-id="eval_aes"><a href='../../AdvancedGgplot/html/Layer.html#method-Layer-eval_aes'><code>AdvancedGgplot::Layer$eval_aes()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeomSmooth-new"></a>}}
\if{latex}{\out{\hypertarget{method-GeomSmooth-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new smooth layer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeomSmooth$new(mapping, method = "loess", span = 0.75)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mapping}}{Aesthetic mappings}

\item{\code{method}}{Smoothing method ("loess" or "lm")}

\item{\code{span}}{Parameter controlling the amount of smoothing}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeomSmooth-render"></a>}}
\if{latex}{\out{\hypertarget{method-GeomSmooth-render}{}}}
\subsection{Method \code{render()}}{
Render the smoothed line and confidence interval
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeomSmooth$render(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data to render}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeomSmooth-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GeomSmooth-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeomSmooth$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
