% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_violin.R
\name{GeomViolin}
\alias{GeomViolin}
\title{Violin Plot Geometry}
\description{
Create a violin plot layer for visualizing distributions
}
\section{Super class}{
\code{\link[AdvancedGgplot:Layer]{AdvancedGgplot::Layer}} -> \code{GeomViolin}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GeomViolin-render}{\code{GeomViolin$render()}}
\item \href{#method-GeomViolin-clone}{\code{GeomViolin$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Layer" data-id="eval_aes"><a href='../../AdvancedGgplot/html/Layer.html#method-Layer-eval_aes'><code>AdvancedGgplot::Layer$eval_aes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Layer" data-id="initialize"><a href='../../AdvancedGgplot/html/Layer.html#method-Layer-initialize'><code>AdvancedGgplot::Layer$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeomViolin-render"></a>}}
\if{latex}{\out{\hypertarget{method-GeomViolin-render}{}}}
\subsection{Method \code{render()}}{
Render violin plots on the plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeomViolin$render(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data frame containing the points to plot}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeomViolin-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GeomViolin-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeomViolin$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
