% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{interactive_plot}
\alias{interactive_plot}
\title{Create an interactive plot}
\usage{
interactive_plot(data = NULL)
}
\arguments{
\item{data}{Data frame containing the data to plot}
}
\description{
Create an interactive plot
}
