% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{interactive_point}
\alias{interactive_point}
\title{Create an interactive point layer}
\usage{
interactive_point(mapping = NULL, hover_style = NULL, click_style = NULL)
}
\arguments{
\item{mapping}{Aesthetic mappings}

\item{hover_style}{Style to apply on hover}

\item{click_style}{Style to apply on click}
}
\description{
Create an interactive point layer
}
