% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{InteractiveLayer}
\alias{InteractiveLayer}
\title{Interactive Layer Class}
\description{
Base class for interactive layers
}
\section{Super class}{
\code{\link[AdvancedGgplot:Layer]{AdvancedGgplot::Layer}} -> \code{InteractiveLayer}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InteractiveLayer-set_hover_style}{\code{InteractiveLayer$set_hover_style()}}
\item \href{#method-InteractiveLayer-set_click_style}{\code{InteractiveLayer$set_click_style()}}
\item \href{#method-InteractiveLayer-clone}{\code{InteractiveLayer$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Layer" data-id="eval_aes"><a href='../../AdvancedGgplot/html/Layer.html#method-Layer-eval_aes'><code>AdvancedGgplot::Layer$eval_aes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Layer" data-id="initialize"><a href='../../AdvancedGgplot/html/Layer.html#method-Layer-initialize'><code>AdvancedGgplot::Layer$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Layer" data-id="render"><a href='../../AdvancedGgplot/html/Layer.html#method-Layer-render'><code>AdvancedGgplot::Layer$render()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InteractiveLayer-set_hover_style"></a>}}
\if{latex}{\out{\hypertarget{method-InteractiveLayer-set_hover_style}{}}}
\subsection{Method \code{set_hover_style()}}{
Set hover style
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InteractiveLayer$set_hover_style(style)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{style}}{List of style properties}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InteractiveLayer-set_click_style"></a>}}
\if{latex}{\out{\hypertarget{method-InteractiveLayer-set_click_style}{}}}
\subsection{Method \code{set_click_style()}}{
Set click style
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InteractiveLayer$set_click_style(style)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{style}}{List of style properties}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InteractiveLayer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InteractiveLayer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InteractiveLayer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
