% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{InteractivePlot}
\alias{InteractivePlot}
\title{Interactive Plot Class}
\description{
Extensions for interactive plot features
}
\section{Super classes}{
\code{\link[AdvancedGgplot:AGPlot]{AdvancedGgplot::AGPlot}} -> \code{\link[AdvancedGgplot:Plot]{AdvancedGgplot::Plot}} -> \code{InteractivePlot}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InteractivePlot-add_tooltip}{\code{InteractivePlot$add_tooltip()}}
\item \href{#method-InteractivePlot-add_click_handler}{\code{InteractivePlot$add_click_handler()}}
\item \href{#method-InteractivePlot-enable_zoom}{\code{InteractivePlot$enable_zoom()}}
\item \href{#method-InteractivePlot-enable_pan}{\code{InteractivePlot$enable_pan()}}
\item \href{#method-InteractivePlot-enable_brush}{\code{InteractivePlot$enable_brush()}}
\item \href{#method-InteractivePlot-as_widget}{\code{InteractivePlot$as_widget()}}
\item \href{#method-InteractivePlot-clone}{\code{InteractivePlot$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="AGPlot" data-id="add_annotation"><a href='../../AdvancedGgplot/html/AGPlot.html#method-AGPlot-add_annotation'><code>AdvancedGgplot::AGPlot$add_annotation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="AGPlot" data-id="add_layer"><a href='../../AdvancedGgplot/html/AGPlot.html#method-AGPlot-add_layer'><code>AdvancedGgplot::AGPlot$add_layer()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="AGPlot" data-id="add_scale"><a href='../../AdvancedGgplot/html/AGPlot.html#method-AGPlot-add_scale'><code>AdvancedGgplot::AGPlot$add_scale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="AGPlot" data-id="collect_legends"><a href='../../AdvancedGgplot/html/AGPlot.html#method-AGPlot-collect_legends'><code>AdvancedGgplot::AGPlot$collect_legends()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="AGPlot" data-id="compute_ranges"><a href='../../AdvancedGgplot/html/AGPlot.html#method-AGPlot-compute_ranges'><code>AdvancedGgplot::AGPlot$compute_ranges()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="AGPlot" data-id="initialize"><a href='../../AdvancedGgplot/html/AGPlot.html#method-AGPlot-initialize'><code>AdvancedGgplot::AGPlot$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="AGPlot" data-id="plot"><a href='../../AdvancedGgplot/html/AGPlot.html#method-AGPlot-plot'><code>AdvancedGgplot::AGPlot$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="AGPlot" data-id="set_coord"><a href='../../AdvancedGgplot/html/AGPlot.html#method-AGPlot-set_coord'><code>AdvancedGgplot::AGPlot$set_coord()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="AGPlot" data-id="set_facet"><a href='../../AdvancedGgplot/html/AGPlot.html#method-AGPlot-set_facet'><code>AdvancedGgplot::AGPlot$set_facet()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="AGPlot" data-id="set_theme"><a href='../../AdvancedGgplot/html/AGPlot.html#method-AGPlot-set_theme'><code>AdvancedGgplot::AGPlot$set_theme()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InteractivePlot-add_tooltip"></a>}}
\if{latex}{\out{\hypertarget{method-InteractivePlot-add_tooltip}{}}}
\subsection{Method \code{add_tooltip()}}{
Enable tooltips for a layer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InteractivePlot$add_tooltip(layer_id, tooltip_func)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{layer_id}}{Layer identifier}

\item{\code{tooltip_func}}{Function to generate tooltip text}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InteractivePlot-add_click_handler"></a>}}
\if{latex}{\out{\hypertarget{method-InteractivePlot-add_click_handler}{}}}
\subsection{Method \code{add_click_handler()}}{
Add click handler for a layer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InteractivePlot$add_click_handler(layer_id, handler_func)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{layer_id}}{Layer identifier}

\item{\code{handler_func}}{Function to handle click events}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InteractivePlot-enable_zoom"></a>}}
\if{latex}{\out{\hypertarget{method-InteractivePlot-enable_zoom}{}}}
\subsection{Method \code{enable_zoom()}}{
Enable zoom functionality
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InteractivePlot$enable_zoom(enable = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{enable}}{Whether to enable zoom}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InteractivePlot-enable_pan"></a>}}
\if{latex}{\out{\hypertarget{method-InteractivePlot-enable_pan}{}}}
\subsection{Method \code{enable_pan()}}{
Enable pan functionality
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InteractivePlot$enable_pan(enable = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{enable}}{Whether to enable pan}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InteractivePlot-enable_brush"></a>}}
\if{latex}{\out{\hypertarget{method-InteractivePlot-enable_brush}{}}}
\subsection{Method \code{enable_brush()}}{
Enable brush selection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InteractivePlot$enable_brush(enable = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{enable}}{Whether to enable brush selection}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InteractivePlot-as_widget"></a>}}
\if{latex}{\out{\hypertarget{method-InteractivePlot-as_widget}{}}}
\subsection{Method \code{as_widget()}}{
Convert plot to HTML widget
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InteractivePlot$as_widget(width = NULL, height = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{width}}{Widget width}

\item{\code{height}}{Widget height}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InteractivePlot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InteractivePlot-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InteractivePlot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
