% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{InteractivePoint}
\alias{InteractivePoint}
\title{Interactive Point Layer}
\description{
Interactive scatter plot points
}
\section{Super classes}{
\code{\link[AdvancedGgplot:Layer]{AdvancedGgplot::Layer}} -> \code{\link[AdvancedGgplot:InteractiveLayer]{AdvancedGgplot::InteractiveLayer}} -> \code{InteractivePoint}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InteractivePoint-render}{\code{InteractivePoint$render()}}
\item \href{#method-InteractivePoint-clone}{\code{InteractivePoint$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Layer" data-id="eval_aes"><a href='../../AdvancedGgplot/html/Layer.html#method-Layer-eval_aes'><code>AdvancedGgplot::Layer$eval_aes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Layer" data-id="initialize"><a href='../../AdvancedGgplot/html/Layer.html#method-Layer-initialize'><code>AdvancedGgplot::Layer$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="InteractiveLayer" data-id="set_click_style"><a href='../../AdvancedGgplot/html/InteractiveLayer.html#method-InteractiveLayer-set_click_style'><code>AdvancedGgplot::InteractiveLayer$set_click_style()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="InteractiveLayer" data-id="set_hover_style"><a href='../../AdvancedGgplot/html/InteractiveLayer.html#method-InteractiveLayer-set_hover_style'><code>AdvancedGgplot::InteractiveLayer$set_hover_style()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InteractivePoint-render"></a>}}
\if{latex}{\out{\hypertarget{method-InteractivePoint-render}{}}}
\subsection{Method \code{render()}}{
Render interactive points
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InteractivePoint$render(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InteractivePoint-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InteractivePoint-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InteractivePoint$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
