% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{Layer}
\alias{Layer}
\title{Layer Class}
\description{
Base class for all layer types
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Layer-new}{\code{Layer$new()}}
\item \href{#method-Layer-eval_aes}{\code{Layer$eval_aes()}}
\item \href{#method-Layer-render}{\code{Layer$render()}}
\item \href{#method-Layer-clone}{\code{Layer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Layer-new"></a>}}
\if{latex}{\out{\hypertarget{method-Layer-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new layer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layer$new(mapping = NULL, stat = NULL, position = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mapping}}{Aesthetic mappings}

\item{\code{stat}}{Statistical transformation to apply}

\item{\code{position}}{Position adjustment to apply}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Layer-eval_aes"></a>}}
\if{latex}{\out{\hypertarget{method-Layer-eval_aes}{}}}
\subsection{Method \code{eval_aes()}}{
Evaluate aesthetics for the layer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layer$eval_aes(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data to evaluate aesthetics for}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Layer-render"></a>}}
\if{latex}{\out{\hypertarget{method-Layer-render}{}}}
\subsection{Method \code{render()}}{
Render the layer (to be implemented by subclasses)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layer$render(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data to render}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Layer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Layer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
