% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{Layout}
\alias{Layout}
\title{Layout Class}
\description{
Manages plot layouts and arrangements
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Layout-new}{\code{Layout$new()}}
\item \href{#method-Layout-add_plot}{\code{Layout$add_plot()}}
\item \href{#method-Layout-calculate_grid}{\code{Layout$calculate_grid()}}
\item \href{#method-Layout-render}{\code{Layout$render()}}
\item \href{#method-Layout-clone}{\code{Layout$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Layout-new"></a>}}
\if{latex}{\out{\hypertarget{method-Layout-new}{}}}
\subsection{Method \code{new()}}{
Initialize layout
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layout$new(ncol = NULL, nrow = NULL, widths = NULL, heights = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ncol}}{Number of columns}

\item{\code{nrow}}{Number of rows}

\item{\code{widths}}{Relative widths of columns}

\item{\code{heights}}{Relative heights of rows}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Layout-add_plot"></a>}}
\if{latex}{\out{\hypertarget{method-Layout-add_plot}{}}}
\subsection{Method \code{add_plot()}}{
Add a plot to the layout
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layout$add_plot(plot)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plot}}{Plot object to add}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Layout-calculate_grid"></a>}}
\if{latex}{\out{\hypertarget{method-Layout-calculate_grid}{}}}
\subsection{Method \code{calculate_grid()}}{
Calculate optimal grid layout
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layout$calculate_grid()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Layout-render"></a>}}
\if{latex}{\out{\hypertarget{method-Layout-render}{}}}
\subsection{Method \code{render()}}{
Render all plots in layout
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layout$render()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Layout-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Layout-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layout$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
