% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{Legend}
\alias{Legend}
\title{Legend Class}
\description{
Handles the creation and rendering of plot legends
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Legend-new}{\code{Legend$new()}}
\item \href{#method-Legend-render}{\code{Legend$render()}}
\item \href{#method-Legend-clone}{\code{Legend$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Legend-new"></a>}}
\if{latex}{\out{\hypertarget{method-Legend-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new legend
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Legend$new(aesthetic, title = NULL, values = NULL, labels = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{aesthetic}}{The aesthetic this legend represents (e.g., "color", "fill")}

\item{\code{title}}{Legend title}

\item{\code{values}}{The values to show in the legend}

\item{\code{labels}}{Labels for the values}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Legend-render"></a>}}
\if{latex}{\out{\hypertarget{method-Legend-render}{}}}
\subsection{Method \code{render()}}{
Render the legend in the current viewport
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Legend$render()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Legend-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Legend-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Legend$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
