% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{palette_colorblind}
\alias{palette_colorblind}
\title{Colorblind-friendly color palette}
\usage{
palette_colorblind()
}
\value{
A ColorPalette object with colorblind-friendly colors
}
\description{
Colorblind-friendly color palette
}
