% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{palette_custom}
\alias{palette_custom}
\title{Create a custom color palette}
\usage{
palette_custom(colors, interpolate = TRUE)
}
\arguments{
\item{colors}{Vector of colors}

\item{interpolate}{Whether to interpolate between colors}
}
\description{
Create a custom color palette
}
