% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{palette_default}
\alias{palette_default}
\title{Default color palette with standard colors}
\usage{
palette_default()
}
\value{
A ColorPalette object with default colors
}
\description{
Default color palette with standard colors
}
