% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{palette_diverging}
\alias{palette_diverging}
\title{Diverging color palette for highlighting differences above/below a middle point}
\usage{
palette_diverging()
}
\value{
A ColorPalette object with diverging colors
}
\description{
Diverging color palette for highlighting differences above/below a middle point
}
