% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{palette_sequential}
\alias{palette_sequential}
\title{Sequential color palette for ordered data}
\usage{
palette_sequential()
}
\value{
A ColorPalette object with sequential colors
}
\description{
Sequential color palette for ordered data
}
