% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Plot}
\alias{Plot}
\title{Plot Class (Legacy)}
\description{
The main plot class that handles all visualizations (deprecated, use AGPlot instead)
}
\section{Super class}{
\code{\link[AdvancedGgplot:AGPlot]{AdvancedGgplot::AGPlot}} -> \code{Plot}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Plot-clone}{\code{Plot$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="AGPlot" data-id="add_annotation"><a href='../../AdvancedGgplot/html/AGPlot.html#method-AGPlot-add_annotation'><code>AdvancedGgplot::AGPlot$add_annotation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="AGPlot" data-id="add_layer"><a href='../../AdvancedGgplot/html/AGPlot.html#method-AGPlot-add_layer'><code>AdvancedGgplot::AGPlot$add_layer()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="AGPlot" data-id="add_scale"><a href='../../AdvancedGgplot/html/AGPlot.html#method-AGPlot-add_scale'><code>AdvancedGgplot::AGPlot$add_scale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="AGPlot" data-id="collect_legends"><a href='../../AdvancedGgplot/html/AGPlot.html#method-AGPlot-collect_legends'><code>AdvancedGgplot::AGPlot$collect_legends()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="AGPlot" data-id="compute_ranges"><a href='../../AdvancedGgplot/html/AGPlot.html#method-AGPlot-compute_ranges'><code>AdvancedGgplot::AGPlot$compute_ranges()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="AGPlot" data-id="initialize"><a href='../../AdvancedGgplot/html/AGPlot.html#method-AGPlot-initialize'><code>AdvancedGgplot::AGPlot$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="AGPlot" data-id="plot"><a href='../../AdvancedGgplot/html/AGPlot.html#method-AGPlot-plot'><code>AdvancedGgplot::AGPlot$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="AGPlot" data-id="set_coord"><a href='../../AdvancedGgplot/html/AGPlot.html#method-AGPlot-set_coord'><code>AdvancedGgplot::AGPlot$set_coord()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="AGPlot" data-id="set_facet"><a href='../../AdvancedGgplot/html/AGPlot.html#method-AGPlot-set_facet'><code>AdvancedGgplot::AGPlot$set_facet()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="AGPlot" data-id="set_theme"><a href='../../AdvancedGgplot/html/AGPlot.html#method-AGPlot-set_theme'><code>AdvancedGgplot::AGPlot$set_theme()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
