% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{plot_grid}
\alias{plot_grid}
\title{Create a grid of plots}
\usage{
plot_grid(plots, ncol = NULL, shared_legend = FALSE, legend_position = "right")
}
\arguments{
\item{plots}{List of plots}

\item{ncol}{Number of columns}

\item{shared_legend}{Whether to add shared legend}

\item{legend_position}{Position of shared legend}
}
\description{
Create a grid of plots
}
