% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{PlotGrid}
\alias{PlotGrid}
\title{Plot Grid Class}
\description{
Create a regular grid of plots
}
\section{Super class}{
\code{\link[AdvancedGgplot:Layout]{AdvancedGgplot::Layout}} -> \code{PlotGrid}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PlotGrid-new}{\code{PlotGrid$new()}}
\item \href{#method-PlotGrid-add_shared_legend}{\code{PlotGrid$add_shared_legend()}}
\item \href{#method-PlotGrid-clone}{\code{PlotGrid$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Layout" data-id="add_plot"><a href='../../AdvancedGgplot/html/Layout.html#method-Layout-add_plot'><code>AdvancedGgplot::Layout$add_plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Layout" data-id="calculate_grid"><a href='../../AdvancedGgplot/html/Layout.html#method-Layout-calculate_grid'><code>AdvancedGgplot::Layout$calculate_grid()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Layout" data-id="render"><a href='../../AdvancedGgplot/html/Layout.html#method-Layout-render'><code>AdvancedGgplot::Layout$render()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotGrid-new"></a>}}
\if{latex}{\out{\hypertarget{method-PlotGrid-new}{}}}
\subsection{Method \code{new()}}{
Initialize plot grid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotGrid$new(plots, ncol = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plots}}{List of plots}

\item{\code{ncol}}{Number of columns}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotGrid-add_shared_legend"></a>}}
\if{latex}{\out{\hypertarget{method-PlotGrid-add_shared_legend}{}}}
\subsection{Method \code{add_shared_legend()}}{
Add shared legend for all plots
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotGrid$add_shared_legend(position = "right")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{position}}{Position of legend ("right", "bottom")}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotGrid-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PlotGrid-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotGrid$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
