% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R, R/position.R
\name{Position}
\alias{Position}
\title{Position Adjustment Class}
\description{
Base class for position adjustments

Base class for position adjustments
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Position-new}{\code{Position$new()}}
\item \href{#method-Position-adjust}{\code{Position$adjust()}}
\item \href{#method-Position-clone}{\code{Position$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Position-new"></a>}}
\if{latex}{\out{\hypertarget{method-Position-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Position$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Position-adjust"></a>}}
\if{latex}{\out{\hypertarget{method-Position-adjust}{}}}
\subsection{Method \code{adjust()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Position$adjust(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Position-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Position-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Position$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Position-new}{\code{Position$new()}}
\item \href{#method-Position-adjust}{\code{Position$adjust()}}
\item \href{#method-Position-clone}{\code{Position$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Position-new"></a>}}
\if{latex}{\out{\hypertarget{method-Position-new}{}}}
\subsection{Method \code{new()}}{
Initialize position adjustment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Position$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Position-adjust"></a>}}
\if{latex}{\out{\hypertarget{method-Position-adjust}{}}}
\subsection{Method \code{adjust()}}{
Adjust positions of data points
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Position$adjust(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data frame containing positions to adjust}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Position-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Position-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Position$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
