% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R, R/position.R
\name{position_dodge}
\alias{position_dodge}
\title{Create dodge position adjustment}
\usage{
position_dodge(width = 0.75)

position_dodge(width = 0.75)
}
\arguments{
\item{width}{Width of the dodging}
}
\description{
Create dodge position adjustment

Create a dodge position adjustment
}
