% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{position_jitter}
\alias{position_jitter}
\title{Create a jitter position adjustment}
\usage{
position_jitter(width = 0.4, height = 0.4)
}
\arguments{
\item{width}{Amount of horizontal jitter}

\item{height}{Amount of vertical jitter}
}
\description{
Create a jitter position adjustment
}
