% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R, R/position.R
\name{position_stack}
\alias{position_stack}
\title{Create stack position adjustment}
\usage{
position_stack()

position_stack()
}
\description{
Create stack position adjustment

Create a stack position adjustment
}
