% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R, R/position.R
\name{PositionDodge}
\alias{PositionDodge}
\title{Dodge Position Adjustment}
\description{
Dodge overlapping elements

Adjust positions by dodging overlapping geometries
}
\section{Super class}{
\code{\link[AdvancedGgplot:Position]{AdvancedGgplot::Position}} -> \code{PositionDodge}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PositionDodge-new}{\code{PositionDodge$new()}}
\item \href{#method-PositionDodge-adjust}{\code{PositionDodge$adjust()}}
\item \href{#method-PositionDodge-clone}{\code{PositionDodge$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PositionDodge-new"></a>}}
\if{latex}{\out{\hypertarget{method-PositionDodge-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PositionDodge$new(width = 0.75)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PositionDodge-adjust"></a>}}
\if{latex}{\out{\hypertarget{method-PositionDodge-adjust}{}}}
\subsection{Method \code{adjust()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PositionDodge$adjust(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PositionDodge-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PositionDodge-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PositionDodge$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[AdvancedGgplot:Position]{AdvancedGgplot::Position}} -> \code{PositionDodge}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PositionDodge-new}{\code{PositionDodge$new()}}
\item \href{#method-PositionDodge-adjust}{\code{PositionDodge$adjust()}}
\item \href{#method-PositionDodge-clone}{\code{PositionDodge$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PositionDodge-new"></a>}}
\if{latex}{\out{\hypertarget{method-PositionDodge-new}{}}}
\subsection{Method \code{new()}}{
Initialize dodge position adjustment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PositionDodge$new(width = 0.75)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{width}}{Width of the dodging}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PositionDodge-adjust"></a>}}
\if{latex}{\out{\hypertarget{method-PositionDodge-adjust}{}}}
\subsection{Method \code{adjust()}}{
Adjust positions by dodging
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PositionDodge$adjust(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data frame containing positions to dodge}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PositionDodge-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PositionDodge-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PositionDodge$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
