% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{PositionJitter}
\alias{PositionJitter}
\title{Jitter Position Adjustment}
\description{
Add random noise to prevent overplotting
}
\section{Super class}{
\code{\link[AdvancedGgplot:Position]{AdvancedGgplot::Position}} -> \code{PositionJitter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PositionJitter-new}{\code{PositionJitter$new()}}
\item \href{#method-PositionJitter-adjust}{\code{PositionJitter$adjust()}}
\item \href{#method-PositionJitter-clone}{\code{PositionJitter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PositionJitter-new"></a>}}
\if{latex}{\out{\hypertarget{method-PositionJitter-new}{}}}
\subsection{Method \code{new()}}{
Initialize jitter position adjustment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PositionJitter$new(width = 0.4, height = 0.4)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{width}}{Amount of horizontal jitter}

\item{\code{height}}{Amount of vertical jitter}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PositionJitter-adjust"></a>}}
\if{latex}{\out{\hypertarget{method-PositionJitter-adjust}{}}}
\subsection{Method \code{adjust()}}{
Adjust positions by adding jitter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PositionJitter$adjust(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data frame containing positions to jitter}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PositionJitter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PositionJitter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PositionJitter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
