% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R, R/position.R
\name{PositionStack}
\alias{PositionStack}
\title{Stack Position Adjustment}
\description{
Stack overlapping elements

Stack overlapping geometries (e.g., for stacked bar charts)
}
\section{Super class}{
\code{\link[AdvancedGgplot:Position]{AdvancedGgplot::Position}} -> \code{PositionStack}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PositionStack-adjust}{\code{PositionStack$adjust()}}
\item \href{#method-PositionStack-clone}{\code{PositionStack$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Position" data-id="initialize"><a href='../../AdvancedGgplot/html/Position.html#method-Position-initialize'><code>AdvancedGgplot::Position$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PositionStack-adjust"></a>}}
\if{latex}{\out{\hypertarget{method-PositionStack-adjust}{}}}
\subsection{Method \code{adjust()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PositionStack$adjust(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PositionStack-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PositionStack-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PositionStack$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[AdvancedGgplot:Position]{AdvancedGgplot::Position}} -> \code{PositionStack}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PositionStack-adjust}{\code{PositionStack$adjust()}}
\item \href{#method-PositionStack-clone}{\code{PositionStack$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Position" data-id="initialize"><a href='../../AdvancedGgplot/html/Position.html#method-Position-initialize'><code>AdvancedGgplot::Position$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PositionStack-adjust"></a>}}
\if{latex}{\out{\hypertarget{method-PositionStack-adjust}{}}}
\subsection{Method \code{adjust()}}{
Adjust positions by stacking
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PositionStack$adjust(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data frame containing positions to stack}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PositionStack-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PositionStack-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PositionStack$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
