% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{Scale}
\alias{Scale}
\title{Scale Class}
\description{
Base class for all scale types
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Scale-new}{\code{Scale$new()}}
\item \href{#method-Scale-transform}{\code{Scale$transform()}}
\item \href{#method-Scale-render_axis}{\code{Scale$render_axis()}}
\item \href{#method-Scale-clone}{\code{Scale$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scale-new"></a>}}
\if{latex}{\out{\hypertarget{method-Scale-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new scale
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scale$new(aesthetic, name = NULL, breaks = NULL, labels = NULL, limits = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{aesthetic}}{The aesthetic this scale applies to (e.g., "x", "y", "color")}

\item{\code{name}}{The name to use for the axis or legend}

\item{\code{breaks}}{Vector of break points}

\item{\code{labels}}{Vector of labels for the breaks}

\item{\code{limits}}{Vector of length 2 giving the range of the scale}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scale-transform"></a>}}
\if{latex}{\out{\hypertarget{method-Scale-transform}{}}}
\subsection{Method \code{transform()}}{
Transform data from data space to plot space
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scale$transform(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Values to transform}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scale-render_axis"></a>}}
\if{latex}{\out{\hypertarget{method-Scale-render_axis}{}}}
\subsection{Method \code{render_axis()}}{
Render the axis or legend for this scale
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scale$render_axis(side = "bottom")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{side}}{The side to render the axis on ("top", "right", "bottom", "left")}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scale-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Scale-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scale$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
