% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{scale_color_continuous}
\alias{scale_color_continuous}
\title{Create a continuous color scale}
\usage{
scale_color_continuous(
  aesthetic,
  palette = palette_sequential(),
  na_color = "#808080"
)
}
\arguments{
\item{aesthetic}{The aesthetic to scale}

\item{palette}{Color palette to use}

\item{na_color}{Color for NA values}
}
\description{
Create a continuous color scale
}
