% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{scale_color_discrete}
\alias{scale_color_discrete}
\title{Create a discrete color scale}
\usage{
scale_color_discrete(
  aesthetic,
  palette = palette_default(),
  na_color = "#808080"
)
}
\arguments{
\item{aesthetic}{The aesthetic to scale}

\item{palette}{Color palette to use}

\item{na_color}{Color for NA values}
}
\description{
Create a discrete color scale
}
