% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_continuous}
\alias{scale_continuous}
\title{Create a continuous scale}
\usage{
scale_continuous(
  aesthetic,
  name = NULL,
  breaks = NULL,
  labels = NULL,
  limits = NULL
)
}
\arguments{
\item{aesthetic}{The aesthetic to scale}

\item{name}{Optional name for the scale}

\item{breaks}{Optional break points}

\item{labels}{Optional labels for breaks}

\item{limits}{Optional limits for the scale}
}
\description{
Create a continuous scale
}
