% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{Stat}
\alias{Stat}
\title{Statistical Transformation Base Class}
\description{
Base class for all statistical transformations
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Stat-new}{\code{Stat$new()}}
\item \href{#method-Stat-compute}{\code{Stat$compute()}}
\item \href{#method-Stat-clone}{\code{Stat$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stat-new"></a>}}
\if{latex}{\out{\hypertarget{method-Stat-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new statistical transformation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stat$new(params = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{Parameters for the transformation}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stat-compute"></a>}}
\if{latex}{\out{\hypertarget{method-Stat-compute}{}}}
\subsection{Method \code{compute()}}{
Compute the statistical transformation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stat$compute(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Input data frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stat-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Stat-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stat$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
