% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{StatBin}
\alias{StatBin}
\title{Binning Statistical Transformation}
\description{
Computes bins for histograms
}
\section{Super class}{
\code{\link[AdvancedGgplot:Stat]{AdvancedGgplot::Stat}} -> \code{StatBin}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StatBin-compute}{\code{StatBin$compute()}}
\item \href{#method-StatBin-clone}{\code{StatBin$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Stat" data-id="initialize"><a href='../../AdvancedGgplot/html/Stat.html#method-Stat-initialize'><code>AdvancedGgplot::Stat$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatBin-compute"></a>}}
\if{latex}{\out{\hypertarget{method-StatBin-compute}{}}}
\subsection{Method \code{compute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatBin$compute(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatBin-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StatBin-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatBin$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
