% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_boxplot.R
\name{StatBoxplot}
\alias{StatBoxplot}
\title{Boxplot Statistics}
\description{
Statistical transformation for boxplot calculations
}
\section{Super class}{
\code{\link[AdvancedGgplot:Stat]{AdvancedGgplot::Stat}} -> \code{StatBoxplot}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StatBoxplot-compute}{\code{StatBoxplot$compute()}}
\item \href{#method-StatBoxplot-clone}{\code{StatBoxplot$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Stat" data-id="initialize"><a href='../../AdvancedGgplot/html/Stat.html#method-Stat-initialize'><code>AdvancedGgplot::Stat$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatBoxplot-compute"></a>}}
\if{latex}{\out{\hypertarget{method-StatBoxplot-compute}{}}}
\subsection{Method \code{compute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatBoxplot$compute(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatBoxplot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StatBoxplot-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatBoxplot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
