% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{StatSmooth}
\alias{StatSmooth}
\title{Smoothing Statistical Transformation}
\description{
Computes smoothed lines and confidence intervals
}
\section{Super class}{
\code{\link[AdvancedGgplot:Stat]{AdvancedGgplot::Stat}} -> \code{StatSmooth}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StatSmooth-compute}{\code{StatSmooth$compute()}}
\item \href{#method-StatSmooth-clone}{\code{StatSmooth$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Stat" data-id="initialize"><a href='../../AdvancedGgplot/html/Stat.html#method-Stat-initialize'><code>AdvancedGgplot::Stat$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatSmooth-compute"></a>}}
\if{latex}{\out{\hypertarget{method-StatSmooth-compute}{}}}
\subsection{Method \code{compute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatSmooth$compute(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatSmooth-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StatSmooth-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatSmooth$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
