% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_violin.R
\name{StatViolin}
\alias{StatViolin}
\title{Violin Plot Statistics}
\description{
Statistical transformation for violin plot density calculations
}
\section{Super class}{
\code{\link[AdvancedGgplot:Stat]{AdvancedGgplot::Stat}} -> \code{StatViolin}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StatViolin-compute}{\code{StatViolin$compute()}}
\item \href{#method-StatViolin-clone}{\code{StatViolin$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="AdvancedGgplot" data-topic="Stat" data-id="initialize"><a href='../../AdvancedGgplot/html/Stat.html#method-Stat-initialize'><code>AdvancedGgplot::Stat$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatViolin-compute"></a>}}
\if{latex}{\out{\hypertarget{method-StatViolin-compute}{}}}
\subsection{Method \code{compute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatViolin$compute(data, scale = "area", width = 0.8)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatViolin-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StatViolin-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatViolin$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
