% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme}
\alias{theme}
\title{Create a new theme}
\usage{
theme(...)
}
\arguments{
\item{...}{Named theme elements}
}
\value{
A Theme object
}
\description{
Creates a new Theme object with the specified elements.
}
\examples{
\dontrun{
my_theme <- theme(
  plot.background = list(fill = "lightyellow"),
  axis.line = list(color = "darkgrey")
)
}
}
