% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_dark}
\alias{theme_dark}
\title{Create a dark theme with dark background}
\usage{
theme_dark()
}
\value{
A AGTheme object with dark styling
}
\description{
Create a dark theme with dark background
}
\examples{
\dontrun{
plot <- AGPlot$new(data)
plot$set_theme(theme_dark())
}
}
