% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_minimal}
\alias{theme_minimal}
\title{Create a minimal theme with light background}
\usage{
theme_minimal()
}
\value{
A AGTheme object with minimal styling
}
\description{
Create a minimal theme with light background
}
\examples{
\dontrun{
plot <- AGPlot$new(data)
plot$set_theme(theme_minimal())
}
}
