% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation.R
\name{TimeSeriesAnimation}
\alias{TimeSeriesAnimation}
\title{Time Series Animation}
\description{
Animate time series data
}
\section{Super class}{
\code{\link[AdvancedGgplot:Animation]{AdvancedGgplot::Animation}} -> \code{TimeSeriesAnimation}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TimeSeriesAnimation-new}{\code{TimeSeriesAnimation$new()}}
\item \href{#method-TimeSeriesAnimation-generate_frames}{\code{TimeSeriesAnimation$generate_frames()}}
\item \href{#method-TimeSeriesAnimation-clone}{\code{TimeSeriesAnimation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeSeriesAnimation-new"></a>}}
\if{latex}{\out{\hypertarget{method-TimeSeriesAnimation-new}{}}}
\subsection{Method \code{new()}}{
Initialize time series animation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeSeriesAnimation$new(time_column, window_size = NULL, duration = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time_column}}{Name of time column}

\item{\code{window_size}}{Size of moving window}

\item{\code{duration}}{Duration in milliseconds}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeSeriesAnimation-generate_frames"></a>}}
\if{latex}{\out{\hypertarget{method-TimeSeriesAnimation-generate_frames}{}}}
\subsection{Method \code{generate_frames()}}{
Generate time series frames
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeSeriesAnimation$generate_frames(data, n_frames = 30)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data frame containing time series}

\item{\code{n_frames}}{Number of frames}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeSeriesAnimation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TimeSeriesAnimation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeSeriesAnimation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
