% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation.R
\name{TransitionAnimation}
\alias{TransitionAnimation}
\title{Transition Animation}
\description{
Animate transitions between states
}
\section{Super class}{
\code{\link[AdvancedGgplot:Animation]{AdvancedGgplot::Animation}} -> \code{TransitionAnimation}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TransitionAnimation-new}{\code{TransitionAnimation$new()}}
\item \href{#method-TransitionAnimation-generate_frames}{\code{TransitionAnimation$generate_frames()}}
\item \href{#method-TransitionAnimation-interpolate}{\code{TransitionAnimation$interpolate()}}
\item \href{#method-TransitionAnimation-clone}{\code{TransitionAnimation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionAnimation-new"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionAnimation-new}{}}}
\subsection{Method \code{new()}}{
Initialize transition animation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionAnimation$new(states, duration = 1000, easing = "linear")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{states}}{List of data frames representing states}

\item{\code{duration}}{Duration in milliseconds}

\item{\code{easing}}{Easing function name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionAnimation-generate_frames"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionAnimation-generate_frames}{}}}
\subsection{Method \code{generate_frames()}}{
Generate transition frames
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionAnimation$generate_frames(data, n_frames = 30)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Initial data frame}

\item{\code{n_frames}}{Number of frames}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionAnimation-interpolate"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionAnimation-interpolate}{}}}
\subsection{Method \code{interpolate()}}{
Interpolate between states
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionAnimation$interpolate(start_state, end_state, alpha)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start_state}}{Starting state}

\item{\code{end_state}}{Ending state}

\item{\code{alpha}}{Interpolation parameter}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TransitionAnimation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TransitionAnimation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionAnimation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
